% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createClinicalDB.R
\name{CreateClinicalDatabase}
\alias{CreateClinicalDatabase}
\title{CreateClinicalDatabase
CreateClinicalDatabase reads PAC and PM clinical reports and creates a
formatted clinical database for usage within INOVIO's Immunology and
Biomarkers group - HPV-301 Program}
\usage{
CreateClinicalDatabase()
}
\description{
CreateClinicalDatabase
CreateClinicalDatabase reads PAC and PM clinical reports and creates a
formatted clinical database for usage within INOVIO's Immunology and
Biomarkers group - HPV-301 Program
}
