% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createVirologyDB.R
\name{CreateVirologyDatabase}
\alias{CreateVirologyDatabase}
\title{CreateVirologyDataBase
CreateVirologyDataBase reads virology Reports reports and creates a formatted
virology database for usage within INOVIO's Immunology and Biomarkers group -
HPV-301 Program}
\usage{
CreateVirologyDatabase()
}
\value{
A data frame with virology data read from the current
 working directory
}
\description{
CreateVirologyDataBase
CreateVirologyDataBase reads virology Reports reports and creates a formatted
virology database for usage within INOVIO's Immunology and Biomarkers group -
HPV-301 Program
}
