% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrganizeClinicalResportsRO.R
\name{OrganizeClinicalReportsRO}
\alias{OrganizeClinicalReportsRO}
\title{OrganizeClinicalReportsRO
OrganizeClinicalReportsRO renames and arranges .pdfs of PAC, PM, and Virology
clinical reports from HPV-301}
\usage{
OrganizeClinicalReportsRO()
}
\value{
Renames and organizes files in the working directory
}
\description{
OrganizeClinicalReportsRO
OrganizeClinicalReportsRO renames and arranges .pdfs of PAC, PM, and Virology
clinical reports from HPV-301
}
